+++ config/config.js
@@ -111,6 +111,11 @@ module.exports = {
     defaultSearchResultsView: "List",
 
     /*
+     * Set to true to highlight search terms in the search result text 
+     */
+    enableSearchHitHighlighting: false,
+
+    /*
      * Set to false if collection objects should be omitted from search results
      */
     showCollectionObjectsInSearchResults: true,
@@ -123,12 +128,14 @@ module.exports = {
     /* 
      * Viewer to play audio files
      * [browser | jwplayer | universalviewer | kaltura]
+     * 'browser' will use <audio> element
      */
     audioPlayer: "universalviewer",
 
     /* 
      * Viewer to display video files
-     * [videojs | jwplayer | universalviewer | kaltura]
+     * [browser | videojs | jwplayer | universalviewer | kaltura]
+     * 'browser' will use <video> element
      */
     videoViewer: "universalviewer",
 
@@ -148,7 +155,9 @@ module.exports = {
     /* 
      * Viewer to display large image files (tiff, jp2)
      * [browser | openseadragon | universalviewer]
+     *
      * "browser" viewer will not display .jp2 images
+     * "browser" option currently will not scale image to browser viewer section
      */
     largeImageViewer: "universalviewer",
 
@@ -315,7 +324,7 @@ module.exports = {
             "type": {
                 "still image": {
                     "source": "iiif",
-                    "cache": true
+                    "cache": false
                 },
                 "audio": {
                     "source": "kaltura",
@@ -327,7 +336,7 @@ module.exports = {
                 },
                 "pdf": {
                     "source": "iiif",
-                    "cache": true
+                    "cache": false
                 }
             }
         }
@@ -352,11 +361,11 @@ module.exports = {
             "type": {
                 "still image": {
                     "source": "repository",
-                    "cache": true,
+                    "cache": false,
                     "file_type": {
                         "jpg": {
                             "source": "iiif",
-                            "cache": true
+                            "cache": false
                         }
                     }
                 },
@@ -370,7 +379,7 @@ module.exports = {
                 },
                 "pdf": {
                     "source": "repository",
-                    "cache": true
+                    "cache": false
                 }
             }
         }
@@ -425,7 +434,7 @@ module.exports = {
         {"label": "Creator", "id": "creator", "field": "creator", "boost": "8"},
         {"label": "Subject", "id": "subject", "field": "f_subjects", "boost": "4"},
         {"label": "Topic", "id": "topic", "field": "display_record.subjects.terms.term", "matchField": "display_record.subjects.terms.type", "matchTerm": "topical"},
-        {"label": "Format", "id": "format", "field": "type", "boost": "3"},
+        {"label": "Format", "id": "type", "field": "type", "boost": "3"},
         {"label": "Description", "id": "description", "field": "abstract", "boost": "6"},
         {"label": "Language", "id": "language", "field": "display_record.t_language.text", "boost": "1"},
         {"label": "Creation Date", "id": "create_date", "field": "display_record.dates.expression", "isNestedType": "true", "matchField": "display_record.dates.label", "matchTerm": "creation"},