+++ search/helper.js
@@ -23,8 +23,10 @@
 
 'use strict';
 
-var config = require('../config/' + process.env.CONFIGURATION_FILE);
-var pluralize = require('pluralize');
+const config = require('../config/' + process.env.CONFIGURATION_FILE);
+const metadataConfig = require('../config/config-metadata-displays');
+const pluralize = require('pluralize');
+const { removeStopwords } = require('stopword');
 
 /**
  * Removes any facets appearing in 'facets' object from the Elastic response object agregations buckets 
@@ -78,11 +80,13 @@ exports.removeEmptyFacetKeys = function(facets) {
  *
  * @return {String} The results label
  */
-exports.getResultsLabel = function(query, facets, bool, field) {
+exports.getSearchTermsLabel = function(query, facets, bool, field) {
   let queryLabel = " ", // One space character is required here (" ")
       appendLabel = "";
+
   if(query && query.length > 0) {
     for(let index in query) {
+
       // Handle special case of a collection field advanced search
       if(field[index] && field[index].toLowerCase() == "collection") {
         continue;
@@ -96,7 +100,7 @@ exports.getResultsLabel = function(query, facets, bool, field) {
       queryLabel += (query[index] + ((index == query.length-1) ? " " : "; "));
     }
   }
-  return queryLabel + appendLabel; 
+  return queryLabel + appendLabel;
 }
 
 /**
@@ -195,24 +199,37 @@ exports.getDateRangeQuery = function(daterange) {
  *
  * @return {Array.<Object>} Array of search fields defined in the configuration
  */
-exports.getSearchFields = function(fieldValue) {
+var getSearchFields = function(field) {
   var fields = [];
 
   // Non-scoped search: Search in all of the fields in the fulltext search
-  if(fieldValue.toLowerCase() == 'all') {
+  if(field.toLowerCase() == 'all') {
     fields = config.searchAllFields;
   }
 
-  // Scoped search: Use the selected type in the search type dropdown
   else {
-    for(var field of config.searchAllFields) {
-      if(field.id.toLowerCase() == fieldValue.toLowerCase()) {
-          fields.push(field);
+    for(var searchField of config.searchAllFields) {
+      if(searchField.id.toLowerCase() == field.toLowerCase()) {
+          fields.push(searchField);
       }
     }
   }
 
   return fields;
+}
+exports.getSearchFields = getSearchFields;
+
+/**
+ * Retrieve an array of all search field ids
+ *
+ * @return {Array.<String>} Array of all search field ids
+ */
+var getSearchAllFieldIds = function() {
+  let ids = [];
+  for(var searchField of config.searchAllFields) {
+    ids.push(searchField.id);
+  }
+  return ids;
 }
 
 /**
@@ -232,7 +249,7 @@ exports.getSearchFields = function(fieldValue) {
  *
  * @return {Array.<queryData>} queryDataArray
  */
-exports.getSearchQueryDataObject = function(queryArray, fieldArray, typeArray, boolArray) {
+exports.getSearchQueryDataObject = function(queryArray, fieldArray, typeArray, boolArray, highlightTerms=false) {
   var queryDataArray = [];
 
   for(var index in queryArray) {
@@ -246,7 +263,8 @@ exports.getSearchQueryDataObject = function(queryArray, fieldArray, typeArray, b
       terms: queryArray[index],
       field: fieldArray[index],
       type: typeArray[index],
-      bool: boolArray[index]
+      bool: boolArray[index],
+      highlight: highlightTerms
     });
   }
 
@@ -388,11 +406,17 @@ exports.getResultSetMinDate = function(facets) {
  *
  * @return {string} - Updated terms
  */
-exports.getSearchTerms = function(queryString) {
-  let terms = "";
-  terms = queryString.toLowerCase().replace(/"/g, '') || "";
-  //terms = singularizeSearchStringTerms(terms);
-  return terms;
+exports.formatSearchTerms = function(queryString) {
+
+  // Remove non-alphanumeric characters, except for control characters '*', '"'
+  queryString = queryString.toLowerCase().replace(/[^a-z0-9*\s"]/gi, '') || "";
+
+  // Remove stop words if the terms are not enclosed in quotation marks
+  if(queryString.match(/"|\*/g) == null) {
+    queryString = removeStopwords(queryString.split(" ")).toString().replace(/,/gi, " ")
+  }
+
+  return queryString;
 }
 
 /**
@@ -412,3 +436,88 @@ var singularizeSearchStringTerms = function(string) {
   return string.trim();
 }
 exports.singularizeSearchStringTerms = singularizeSearchStringTerms;
+
+/**
+ * 
+ *
+ * @param {Object}
+ *
+ */
+exports.addSearchTermHighlights = function(queryData, resultObject) {
+  let highlightTerms = [],
+      highlightSearchFields = [];
+
+  let terms, quotedTerms, remainingTerms;
+  for(var index in queryData) {
+    if(queryData[index].highlight) {
+      terms = queryData[index].terms || "";;
+      quotedTerms = [];
+
+      if(queryData[index].type == "is") {
+        quotedTerms.push(terms);
+      }
+      else {
+        quotedTerms = terms.match(/"[a-zA-Z0-9]*[^"]+[a-zA-Z0-9]*"/g);
+      }
+
+      if(quotedTerms && quotedTerms.length > 0) {
+        for(let quotedTerm of quotedTerms) {
+          highlightTerms.push(quotedTerm.replace(/"/g, ""));
+          terms = terms.replace(quotedTerm, "");
+        }
+
+        if(terms.length > 2) {
+          highlightTerms = highlightTerms.concat(terms.split(" ").filter(i => i));
+        }
+      }
+      else if(terms.match(/[a-zA-Z0-9]+/g)) { // only highlight alphanumeric terms
+        highlightTerms = highlightTerms.concat(terms.trim().split(" "));
+      }
+
+      if(queryData[index].field.toLowerCase() == "all") {
+        highlightSearchFields = highlightSearchFields.concat(getSearchAllFieldIds());
+      }
+      else {
+        highlightSearchFields.push(queryData[index].field);
+      }
+    }
+  }
+
+  if(highlightSearchFields.length > 0) {
+    let metadataDisplay;
+    if(resultObject.objectType == "collection") {
+      metadataDisplay = metadataConfig.resultsDisplay["collection"] || {};
+    }
+    else {
+      metadataDisplay = metadataConfig.resultsDisplay["default"] || {};
+    }
+
+    let metadataFieldArray = [];
+    for(let result of resultObject) {
+      for(let index in highlightTerms) {
+
+        if(highlightSearchFields.includes("title")) {
+          result.title = result.title.replace(new RegExp(highlightTerms[index], 'gi'), `<span class="text-highlight">${highlightTerms[index]}</span>`);
+        }
+
+        for(let key in result.metadata) {
+          let searchFieldId = metadataDisplay[key].searchFieldID || "",
+              isHighlighted = highlightSearchFields.includes(metadataDisplay[key].searchFieldId);
+
+          if(typeof result.metadata[key] == "string" && isHighlighted) {
+            result.metadata[key] = result.metadata[key].replace(new RegExp(highlightTerms[index], 'gi'), `<span class="text-highlight">${highlightTerms[index]}</span>`);
+          }
+          else if(typeof result.metadata[key] == "object" && isHighlighted) {
+            metadataFieldArray = [];
+            for(var value of result.metadata[key]) {
+              metadataFieldArray.push(value.replace(new RegExp(highlightTerms[index], 'gi'), `<span class="text-highlight">${highlightTerms[index]}</span>`));
+            }
+            result.metadata[key] = metadataFieldArray;
+          }
+        }
+      }
+    }
+  }
+
+  return resultObject;
+}