+++ changelog.md
@@ -1,4 +1,4 @@
-## Releases
+git ## Releases
 * v1.0.0-beta [release]() [notes]()
 
 ### v1.1.1 6/10/2021
@@ -59,3 +59,16 @@ Fixed bug in compound object part file download
 ### v1.4.1
 
 Datastream module upgrade complete
+
+### v1.4.2
+
+Added search result hit highlighting to add highlight to search terms appearing in current search field(s) in the search results view item data
+Applied stop words filter to search terms in the search service, to prevent hits on single stop words and to display terms without stop words in the results label
+Removed stop words from the search results view search terms label
+Added configuration setting to disable search hit highlighting
+Updated search term formatter to not remove stop words from terms enclosed in quotation marks #144
+Added configuration setting to enable/disable search term highlighting #144
+
+### v1.4.21
+
+ Fixed bug that prevented 'contains' queries from being combined with 'is' queries #144